requirements.txt
streamlit
pandas
scikit-learn
xgboost
joblib
psycopg2-binary
sqlalchemy
python-dotenv